iio: adc: ad7768-1: use new spi-engine implementation
Add  support for new spi-engine implementation on ad7768-1 driver,
optimizing offload transfers by reducing CS delay and minimizing
the number of commands in the offload FIFO.

Without the otimization, there's not enough time between the end of a
transfer and the start of the next one, reducing the actual sample rate.

Adjust 'tx_data' buffer handling in SPI register read and write functions
to ensure endianness consistency with the new spi-engine.

Signed-off-by: Jonathan Santos <Jonathan.Santos@analog.com>