iio: adc: adrv9002: update mcs implementation
There are new bits being supported in HW (hdl) for supporting MCS. Now
we have a way of stating if the MCS pulses are being generated
externally or internally on the FPGA. On top of that we can also have
external vs internal trigger to start those pulses. When both the pulses
and trigger are internal, we need to actual tell the FPGA to start sending
those pules.

Since the above depends on HW arrangement, two new DT properties are
being added to specify if the pulses are internal or external. Same goes
for the trigger of those pulses.

Another thing to note is that now we have a mux for selecting MCS or
DATA running (data is "business" as usual) so that we need to select
MCS during MCS and make sure we leave that mode after we done so that we
can start sending/receiving data.

Signed-off-by: Nuno Sa <nuno.sa@analog.com>