iio; adc: adrv9002: use define for init cals timeout
Since we call the initial calibrations API from multiple places, add a
define for the timeout. This is in preparation of supporting RX Port
Switching where we need to change the timeout and so we only need to
change in one place.

Signed-off-by: Nuno Sá <nuno.sa@analog.com>
(cherry picked from commit 74254321c97bd353881438ee0006beb8e463b8b6)