spi: rename spi_engine to legacy_spi_engine
To prepare for backporting the SPI offload support from upstream, we
need to move the existing SPI engine code to a different file to avoid
conflicts with the new code.

Also, rename the header file and old APIs so that it is clear that they
should not be used in new code.

Signed-off-by: David Lechner <dlechner@baylibre.com>
(cherry picked from commit 58edc6c13a5ed03db4ac0b1194133249845244ec)